Sprint 12: финальные правки, глобальная переделка архитектуры,добавление конструкторов для тестов,рефакторинг

+++ src/main/java/ru/yandex/practicum/filmorate/dal/BaseRepository.java
@@ -55,7 +55,7 @@ public class BaseRepository<T> {
             }, keyHolder);
 
             if (keyHolder.getKeys().size() > 1) {
-                return (long) 1;
+                return 1;
             }
             Number id = keyHolder.getKey();
             if (id != null) {
